refactor: poultryService error no
