+++ Assets/furious-poultry/application/furious-poultry.application.asmdef
@@ -1,7 +1,9 @@
 {
-    "name": "furious-poultry.domain",
+    "name": "furious-poultry.application",
     "rootNamespace": "com.github.UnityWorkshop",
-    "references": [],
+    "references": [
+        "GUID:af9f40918dfcb2846b5cb8e1992888e7"
+    ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
