+++ Assets/furious-poultry/application/furious-poultry.application.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5644c0e3e2d3a244a852df9443333073
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
