+++ Assets/furious-poultry/application/interfaces/IDestructionProvider.cs
@@ -1,4 +1,4 @@
-namespace com.github.UnityWorkshop.furious_poultry.domain
+namespace com.github.UnityWorkshop.furious_poultry.application.interfaces
 {
     public interface IDestructionProvider
     {
