+++ Assets/furious-poultry/application/services/PoultryService.cs.meta
