+++ Assets/furious-poultry/domain.tests/ClampableHealthTest.cs.meta
