+++ Assets/furious-poultry/domain.tests/ClampableIndexTest.cs.meta
