+++ Assets/furious-poultry/domain.tests/WarthogTest.cs
@@ -1,7 +1,7 @@
 ﻿using com.github.UnityWorkshop.furious_poultry.domain;
 using NUnit.Framework;
 
-namespace furious_poultry.domain.tests
+namespace furious_poultry.unity.definition.domain.tests
 {
     public class WarthogTest
     {
