+++ Assets/furious-poultry/domain/Poultry.cs
@@ -10,18 +10,16 @@ namespace com.github.UnityWorkshop.furious_poultry.domain
         public bool IsOnGround { get; set; }
         public bool HasCollided { get; private set; }
         public float DecayTickDamage { get; }
-        private IDestructionProvider _destructionProvider;
         private ClampableHealth _healthClamp;
         
 
-        public Poultry(float decayTickDamage, float damage, float health, bool isOnGround, bool hasCollided, )
+        public Poultry(float decayTickDamage, float damage, float health, bool isOnGround, bool hasCollided)
         {
             this.DecayTickDamage = decayTickDamage;
             this.Damage = damage;
             this.Health = health;
             this.IsOnGround = isOnGround;
             this.HasCollided = hasCollided;
-            _destructionProvider = destructionProvider;
             _healthClamp = new ClampableHealth(Health, 0, health);
         }
 
@@ -41,9 +39,7 @@ namespace com.github.UnityWorkshop.furious_poultry.domain
             Harm(Damage);
             HasCollided = true;
         }
-
         
-
         public void CollidedWithNotGround()
         {
             HasCollided = true;
