+++ Assets/furious-poultry/unity/definition/PoultryDefinition.cs
@@ -1,8 +1,7 @@
 using System;
 using com.github.UnityWorkshop.furious_poultry.domain;
-using com.github.UnityWorkshop.furious_poultry.unity;
 
-namespace furious_poultry.unity
+namespace com.github.UnityWorkshop.furious_poultry.unity.definition
 {   
     [Serializable]
     public class PoultryDefinition
@@ -13,9 +12,9 @@ namespace furious_poultry.unity
         public bool hasCollided;
         public float decayTickDamage = 1;
 
-        public Poultry ToPoultry(IDestructionProvider destructionProvider)
+        public Poultry ToPoultry()
         {
-            return new Poultry(decayTickDamage, damage, health, isOnGround, hasCollided, destructionProvider);
+            return new Poultry(decayTickDamage, damage, health, isOnGround, hasCollided);
         }
 
     }
