+++ Assets/furious-poultry/unity/furious-poultry.unity.asmdef
@@ -2,7 +2,8 @@
     "name": "furious-poultry.unity",
     "rootNamespace": "com.github.UnityWorkshop",
     "references": [
-        "GUID:af9f40918dfcb2846b5cb8e1992888e7"
+        "GUID:af9f40918dfcb2846b5cb8e1992888e7",
+        "GUID:5644c0e3e2d3a244a852df9443333073"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
